(PC-7757) split the offer component from its body
